input clears when added and scrolls to bottom of page
